08 Unit Test in Android  - Unit testing is core of TDD. However, Android is not a hospitable environment for Unit testing. Why?  - Try:     - make sure you select Unit Test artifact in Build Variants panel. This will make unit tests to be added to test package     - add JsonAdapterTest class in core package. Extend TestCase for basic test functionality     - add setUp() and tearDown() methods     - create test method/s with "test" prefix in name (see e.g. testJsonAdapterStores())
Upon running test you will get that "java.lang.RuntimeException" saying that "method length in org.json.JSONArray not mocked. See http://g.co/androidstudio/not-mocked for details". The website propose will mislead you as it will suggest to add default method behavior (e.g. return 0) which will still not give expected results.

The problem here is that the android.jar supplied with the SDK is stubbed out with no implementation code. The solution that seems to be expected is that you should run your unit tests on the emulator or a real phone, or try to provide the test case with missing real jar. Thus we tried to include json.jar and to include it in testCompile, but however this lead us only to the next missing library and thus we will run unit tests on android device / emulator as well.