09 Developing a purchase flow (SERIOUS commit)  - create in app products in google play and get products SKUs  - start purchase logic in new click listener on buy button (onBuyMapButtonClicked)  - add listener on purchase finish event (mPurchaseFinishedListener)  - develop method to create and verify developer payload  - add private fields when needed  - add your SKU-s and add Google test SKU-s for testing purposes  - change onViewCreated to add code to enable/disable components and to set click listener on buy button  - add and call setupIabHelper() method  - add optional listeners for inventory query. Not that it would be much better if inventory query would be executed in MainActivity.  - for testing purposes add Item consumption call and listener
Note 01: InApp purchase is still not testable on your device. Wait for next commit instruction :)
Note 02: The purchase flow presumes that application with billing permission is already build for release and uploaded to store in order to have in app products enabled.
Note 03: When testing, Security class and verifyPurchase method should be temporarily changed as Google does not sign test purchases. More details can be found here: http://stackoverflow.com/questions/19732025/android-in-app-billing-purchase-verification-failed
Note 04: The listener for purchase finish will not into isolated storage that item is bought. However, the main drawer UI is not updated :(
Note 05:  In terms of simplicity, our implementation is situated in a single file (BuyMapFragment). For Homework, separate the purchase flow and logic to new classes that can be reusable when developing infrastructure for other iab products.